pyproject.toml
[tool.poetry]
name = "tetr-io-rank-checker"
version = "0.1.0"
description = ""
authors = ["timelessnesses <mooping3roblox@gmail.com>"]
readme = "README.md"
packages = [{include = "tetr"}]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.1.1"
asyncpg = "^0.27.0"
psycopg2 = "^2.9.5"
humanfriendly = "^10.0"
types-humanfriendly = "^10.0.1.7"
jishaku = "^2.5.1"
bigjson = "^1.0.9"
requests = "^2.28.2"
psycopg2-binary = {version = "^2.9.5", platform = "linux"}
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
